Move books form and lists to containers
